@@ -1,272 +1,250 @@
+import React, { useState } from "react";
 import Link from "next/link";
-import React, { useEffect, useState } from "react";
-import Logo from "./Logo";
-import { useRouter } from "next/router";
-import {
-  DiscordIcon,
-  FacebookIcon,
-  GithubIcon,
-  GmailIcon,
-  LinkedInIcon,
-  TwitterIcon,
-} from "./Icons";
+import Image from "next/image";
+import { usePathname } from "next/navigation";
 import { motion } from "framer-motion";
-import useThemeSwitcher from "./hooks/useThemeSwitcher";
-const CustomLink = ({ href, title, className = "" }) => {
-  const router = useRouter();
-  // console.log(router);
-  return (
-    <Link href={href} className={`${className} relative group`}>
-      {title}
-      <span
-        className={`h-[1px] inline-block bg-dark absolute -bottom-0.5 left-0 group-hover:w-full 
-        transition-[width] ease duration-300
-        ${router.asPath === href ? "w-full" : "w-0"} dark:bg-light`}
-      >
-        &nbsp;
-      </span>
-    </Link>
-  );
-};
-const CustomMobileLink = ({ href, title, className = "", toggle }) => {
-  const router = useRouter();
+import { Menu, X, Linkedin, Facebook } from "lucide-react";
 
-  const handleClick = () => {
-    toggle();
-    router.push(href);
-  };
-  return (
-    <button
-      href={href}
-      className={`${className} relative group text-light dark:text-dark my-2`}
-      onClick={handleClick}
-    >
-      {title}
-      <span
-        className={`h-[1px] inline-block bg-light absolute -bottom-0.5 left-0 group-hover:w-full 
-        transition-[width] ease duration-300
-        ${router.asPath === href ? "w-full" : "w-0"} dark:bg-dark`}
-      >
-        &nbsp;
-      </span>
-    </button>
-  );
-};
-const thresholdWidthForMobile = 1023;
 const NavBar = () => {
-  const [mode, setMode] = useThemeSwitcher();
-  const [isOpen, setIsopen] = useState(false);
-  const [isMobile, setIsMobile] = useState(false);
-  const handleClick = () => {
-    setIsopen(!isOpen);
+  const [isOpen, setIsOpen] = useState(false);
+  const pathname = usePathname();
+
+  const navItems = [
+    { name: "Home", path: "/" },
+    { name: "Experience", path: "/experience" },
+    { name: "Projects", path: "/projects" },
+    { name: "Achievements", path: "/achievements" },
+  ];
+
+  const socialLinks = [
+    {
+      icon: "/images/svgs/github.svg",
+      href: "https://github.com/ChyYasir",
+      color: "hover:text-purple-400",
+      label: "GitHub",
+    },
+    {
+      // icon: Linkedin,
+      icon: "/images/svgs/linkedin-2.svg",
+      href: "https://www.linkedin.com/in/yasir-rahman-chy/",
+      color: "hover:text-blue-400",
+      label: "LinkedIn",
+    },
+    {
+      icon: "/images/svgs/facebook.svg",
+      href: "https://www.facebook.com/profile.php?id=100042767077083",
+      color: "hover:text-cyan-400",
+      label: "Facebook",
+    },
+  ];
+
+  const isActive = (path) => {
+    if (path === "/" && pathname !== "/") return false;
+    return pathname.startsWith(path);
   };
-  useEffect(() => {
-    const handleResize = () => {
-      setIsMobile(window.innerWidth <= thresholdWidthForMobile);
-    };
 
-    handleResize(); // Initial check
-    window.addEventListener("resize", handleResize);
-
-    return () => {
-      window.removeEventListener("resize", handleResize);
-    };
-  }, []);
   return (
-    <>
-      <header
-        className="w-full px-32 py-8 font-medium flex items-center justify-between dark:text-light
-      relative z-10 lg:px-16 md:px-12 sm:px-8"
-      >
-        <button
-          className="lg:flex flex-col justify-center items-center hidden"
-          onClick={handleClick}
-        >
-          <span
-            className={`bg-dark dark:bg-light transition-all duration-300 ease-out block h-0.5 w-6 rounded-sm  ${
-              isOpen ? "rotate-45 translate-y-1" : "-translate-y-0.5"
-            }`}
-          ></span>
-          <span
-            className={`bg-dark dark:bg-light transition-all duration-300 ease-out block h-0.5 w-6 rounded-sm my-0.5 ${
-              isOpen ? "opacity-0" : "opacity-100"
-            }`}
-          ></span>
-          <span
-            className={`bg-dark dark:bg-light transition-all duration-300 ease-out block h-0.5 w-6 rounded-sm ${
-              isOpen ? "-rotate-45 -translate-y-1" : "translate-y-0.5"
-            }`}
-          ></span>
-        </button>
-
-        <div className="w-full flex justify-between items-center lg:hidden">
-          <nav>
-            <CustomLink href="/" title="Home" className="mr-4" />
-            <CustomLink href="/about" title="About" className="mx-4 " />
-            <CustomLink href="/projects" title="Projects" className="mx-4" />
-            <CustomLink
-              href="/achievements"
-              title="Achievements"
-              className="mx-4"
-            />
-          </nav>
-
-          <nav className="flex items-center justify-center flex-wrap">
-            <motion.a
-              href="https://www.linkedin.com/in/yasir-rahman-chy/"
-              target={"_blank"}
-              className="w-8 mr-3"
-              whileHover={{ y: -2 }}
-              whileTap={{ scale: 0.9 }}
-            >
-              <LinkedInIcon />
-            </motion.a>
-            <motion.a
-              href="https://github.com/ChyYasir"
-              target={"_blank"}
-              className="w-7 mx-3"
-              whileHover={{ y: -2 }}
-              whileTap={{ scale: 0.9 }}
-            >
-              <GithubIcon />
-            </motion.a>
-            <motion.a
-              href="https://discord.com/users/758387316103446599"
-              target={"_blank"}
-              className="w-7 mx-3"
-              whileHover={{ y: -2 }}
-              whileTap={{ scale: 0.9 }}
-            >
-              <DiscordIcon />
-            </motion.a>
-
-            <motion.a
-              href="https://www.facebook.com/profile.php?id=100042767077083"
-              target={"_blank"}
-              className="w-7 mx-3"
-              whileHover={{ y: -2 }}
-              whileTap={{ scale: 0.9 }}
-            >
-              <FacebookIcon />
-            </motion.a>
-            <motion.a
-              href={"mailto:chyyasir2000@gmail.com"}
-              target={"_blank"}
-              className="w-7 mx-3"
-              whileHover={{ y: -2 }}
-              whileTap={{ scale: 0.9 }}
-            >
-              <GmailIcon />
-            </motion.a>
-            <button
-              onClick={() => setMode(mode === "dark" ? "light" : "dark")}
-              className={`w-6 ml-5 text-2xl flex items-center justify-center rounded-full p-1
-            `}
-            >
-              {mode === "dark" ? "☀️" : "🌙"}
-            </button>
-          </nav>
-        </div>
-
-        {/* for mobile  */}
-
-        {isOpen && isMobile ? (
+    <motion.nav
+      initial={{ y: -100 }}
+      animate={{ y: 0 }}
+      className="fixed w-full top-0 z-50 bg-black/80 backdrop-blur-md border-b border-green-500/20"
+    >
+      <div className="max-w-7xl mx-auto px-4">
+        <div className="flex items-center justify-between h-20">
+          {/* Logo */}
           <motion.div
-            className="min-w-[70vw] flex flex-col justify-between items-center fixed z-30 top-1/2 left-1/2 
-        -translate-x-1/2 -translate-y-1/2 bg-dark/90 dark:bg-light/75 rounded-lg backdrop-blur-md py-32"
-            initial={{ scale: 0, opacity: 0, x: "-50%", y: "-50%" }}
-            animate={{ scale: 1, opacity: 100 }}
+            className="text-2xl md:text-3xl font-mono relative group"
+            whileHover={{ scale: 1.05 }}
           >
-            <nav className="flex flex-col items-center justify-center">
-              <CustomMobileLink
-                href="/"
-                title="Home"
-                className=""
-                toggle={handleClick}
-              />
-              <CustomMobileLink
-                href="/about"
-                title="About"
-                className=""
-                toggle={handleClick}
-              />
-              <CustomMobileLink
-                href="/projects"
-                title="Projects"
-                className=""
-                toggle={handleClick}
-              />
-              <CustomMobileLink
-                href="/achievements"
-                title="Achievements"
-                className=""
-                toggle={handleClick}
-              />
-            </nav>
-
-            <nav className="flex items-center justify-center flex-wrap mt-2">
-              <motion.a
-                href="https://www.linkedin.com/in/yasir-rahman-chy/"
-                target={"_blank"}
-                className="w-6 mr-2"
-                whileHover={{ y: -2 }}
-                whileTap={{ scale: 0.9 }}
-              >
-                <LinkedInIcon />
-              </motion.a>
-              <motion.a
-                href="https://github.com/ChyYasir"
-                target={"_blank"}
-                className="w-6 mx-2 dark:text-dark text-light"
-                whileHover={{ y: -2 }}
-                whileTap={{ scale: 0.9 }}
-              >
-                <GithubIcon />
-              </motion.a>
-              <motion.a
-                href="https://discord.com/users/758387316103446599"
-                target={"_blank"}
-                className="w-6 mx-2"
-                whileHover={{ y: -2 }}
-                whileTap={{ scale: 0.9 }}
-              >
-                <DiscordIcon />
-              </motion.a>
-
-              <motion.a
-                href="https://www.facebook.com/profile.php?id=100042767077083"
-                target={"_blank"}
-                className="w-6 mx-2"
-                whileHover={{ y: -2 }}
-                whileTap={{ scale: 0.9 }}
-              >
-                <FacebookIcon />
-              </motion.a>
-              <motion.a
-                href={"mailto:chyyasir2000@gmail.com"}
-                target={"_blank"}
-                className="w-6 mx-2"
-                whileHover={{ y: -2 }}
-                whileTap={{ scale: 0.9 }}
-              >
-                <GmailIcon />
-              </motion.a>
-              <button
-                onClick={() => setMode(mode === "dark" ? "light" : "dark")}
-                className={`w-6 ml-2 text-2xl flex items-center justify-center rounded-full p-1
-            `}
-              >
-                {mode === "dark" ? "☀️" : "🌙"}
-              </button>
-            </nav>
+            <Link href="/">
+              <div className="cursor-pointer relative px-3 py-2">
+                <div className="absolute -inset-2 bg-gradient-to-r from-green-500/20 via-blue-500/20 to-purple-500/20 rounded-lg blur opacity-0 group-hover:opacity-100 transition duration-300" />
+                <div className="relative">
+                  <span className="text-gray-400">&lt;</span>
+                  <span className="text-green-400">YR</span>
+                  <span className="text-gray-400">/&gt;</span>
+                </div>
+              </div>
+            </Link>
           </motion.div>
-        ) : null}
 
-        <div className="absolute left-[50%] top-2 translate-x-[-50%]">
-          <Logo />
+          {/* Desktop Navigation - Centered */}
+          <div className="hidden md:flex items-center justify-center flex-1 mx-4">
+            <div className="flex items-center space-x-3">
+              {navItems.map((item, index) => {
+                const active = isActive(item.path);
+                return (
+                  <Link key={index} href={item.path}>
+                    <motion.div
+                      className="relative group"
+                      whileHover={{ scale: 1.05 }}
+                    >
+                      <div className="px-5 py-3">
+                        <div
+                          className={`absolute inset-0 rounded-lg transition-all duration-300
+                          ${
+                            active
+                              ? "bg-gradient-to-r from-green-500/20 via-green-500/10 to-green-500/20 border border-green-500/30"
+                              : "bg-transparent group-hover:bg-gradient-to-r from-green-500/10 via-green-500/5 to-green-500/10"
+                          }`}
+                        />
+                        <div className="relative flex items-center">
+                          <span
+                            className={`font-mono text-lg tracking-wide transition-all duration-300
+                            ${
+                              active
+                                ? "text-green-400"
+                                : "text-gray-400 group-hover:text-green-400"
+                            }`}
+                          >
+                            {item.name}
+                          </span>
+                          {active && (
+                            <motion.div
+                              layoutId="activeIndicator"
+                              className="absolute -bottom-1 left-0 right-0 h-0.5 bg-gradient-to-r from-green-400/50 via-green-400 to-green-400/50"
+                            />
+                          )}
+                        </div>
+                      </div>
+                    </motion.div>
+                  </Link>
+                );
+              })}
+            </div>
+          </div>
+
+          {/* Social Links - Desktop */}
+          <div className="hidden md:flex items-center space-x-5">
+            {socialLinks.map((social, index) => (
+              <motion.a
+                key={index}
+                href={social.href}
+                target="_blank"
+                rel="noopener noreferrer"
+                className={`relative group ${social.color} transition-colors duration-300`}
+                whileHover={{ scale: 1.1 }}
+                title={social.label}
+              >
+                <div
+                  className="absolute -inset-2 bg-gradient-to-r from-green-500/20 via-blue-500/20 to-purple-500/20 
+                              rounded-lg blur opacity-0 group-hover:opacity-100 transition duration-300"
+                />
+                <div className="relative text-gray-400 group-hover:text-inherit">
+                  {typeof social.icon === "string" ? (
+                    <Image
+                      src={social.icon}
+                      alt={social.label}
+                      width={28}
+                      height={28}
+                      className="transform group-hover:rotate-12 transition-transform duration-300"
+                    />
+                  ) : (
+                    <social.icon
+                      size={28}
+                      className="transform group-hover:rotate-12 transition-transform duration-300"
+                    />
+                  )}
+                </div>
+              </motion.a>
+            ))}
+          </div>
+
+          {/* Mobile Menu Button */}
+          <div className="md:hidden">
+            <motion.button
+              whileHover={{ scale: 1.1 }}
+              whileTap={{ scale: 0.9 }}
+              onClick={() => setIsOpen(!isOpen)}
+              className="relative group p-2"
+            >
+              <div className="absolute inset-0 bg-gradient-to-r from-green-500/20 to-blue-500/20 rounded-lg blur opacity-0 group-hover:opacity-100 transition duration-200" />
+              <div className="relative text-green-400 hover:text-green-300 transition-colors">
+                {isOpen ? <X size={28} /> : <Menu size={28} />}
+              </div>
+            </motion.button>
+          </div>
         </div>
-      </header>
-    </>
+
+        {/* Mobile Navigation */}
+        {isOpen && (
+          <motion.div
+            initial={{ opacity: 0, y: -20 }}
+            animate={{ opacity: 1, y: 0 }}
+            exit={{ opacity: 0, y: -20 }}
+            className="md:hidden"
+          >
+            <div className="px-3 pt-3 pb-4 space-y-2 bg-black/90 rounded-lg border border-green-500/20 mb-2">
+              {navItems.map((item, index) => {
+                const active = isActive(item.path);
+                return (
+                  <Link key={index} href={item.path}>
+                    <motion.div
+                      className="relative overflow-hidden"
+                      onClick={() => setIsOpen(false)}
+                    >
+                      <div
+                        className={`px-4 py-3 rounded-lg relative
+                        ${
+                          active
+                            ? "bg-gradient-to-r from-green-500/20 via-green-500/10 to-green-500/20 border border-green-500/30"
+                            : "hover:bg-gradient-to-r hover:from-green-500/10 hover:via-green-500/5 hover:to-green-500/10"
+                        }`}
+                      >
+                        <div className="flex items-center space-x-3">
+                          <span
+                            className={`font-mono text-lg transition-colors duration-300
+                            ${active ? "text-green-400" : "text-gray-400"}`}
+                          >
+                            {item.name}
+                          </span>
+                        </div>
+                      </div>
+                    </motion.div>
+                  </Link>
+                );
+              })}
+
+              {/* Social Links - Mobile */}
+              <div className="flex items-center justify-around pt-4 border-t border-green-500/20 mt-4">
+                {socialLinks.map((social, index) => (
+                  <motion.a
+                    key={index}
+                    href={social.href}
+                    target="_blank"
+                    rel="noopener noreferrer"
+                    className={`text-gray-400 ${social.color} p-2 relative group`}
+                    whileHover={{ scale: 1.15 }}
+                    title={social.label}
+                  >
+                    <div
+                      className="absolute -inset-2 bg-gradient-to-r from-green-500/20 via-blue-500/20 to-purple-500/20 
+                                  rounded-lg blur opacity-0 group-hover:opacity-100 transition duration-300"
+                    />
+                    <div className="relative">
+                      {typeof social.icon === "string" ? (
+                        <Image
+                          src={social.icon}
+                          alt={social.label}
+                          width={24}
+                          height={24}
+                          className="transform group-hover:rotate-12 transition-transform duration-300"
+                        />
+                      ) : (
+                        <social.icon
+                          size={24}
+                          className="transform group-hover:rotate-12 transition-transform duration-300"
+                        />
+                      )}
+                    </div>
+                  </motion.a>
+                ))}
+              </div>
+            </div>
+          </motion.div>
+        )}
+      </div>
+    </motion.nav>
   );
 };
 
