@@ -1,43 +1,27 @@
 import React from "react";
-import Layout from "./Layout";
-import Link from "next/link";
+import { motion } from "framer-motion";
 
-const Footer = () => {
-  return (
-    <footer
-      className="w-full border-solid border-t-2 border-dark font-medium text-lg dark:text-light
-    sm:text-base"
-    >
-      <Layout
-        className="py-8 flex items-center justify-between 
-      lg:flex-col lg:py-6 "
+const Footer = () => (
+  <footer className="bg-black/80 border-t border-green-500/30 py-4">
+    <div className="max-w-4xl mx-auto text-center">
+      <motion.p
+        className="text-green-400 font-mono text-sm"
+        initial={{ opacity: 0 }}
+        whileInView={{ opacity: 1 }}
+        transition={{ duration: 0.8 }}
+        viewport={{ once: true }}
       >
-        <span>© {new Date().getFullYear()}. All Rights Reserved.</span>
-        <div className="flex items-center lg:py-2">
-          Made with ❤️ by &nbsp;
-          <Link
-            href={"https://www.linkedin.com/in/yasir-rahman-chy/"}
-            className="underline underline-offset-2"
-            target="_blank"
-          >
-            Yasir Rahman
-          </Link>{" "}
-        </div>
-        <div className="flex flex-col text-center">
-          <span className="mb-2">
-            If you have any queries or project in mind
-          </span>
-          <Link
-            href={"mailto:chyyasir2000@gmail.com"}
-            target="_blank"
-            className="underline underline-offset-2 "
-          >
-            Say, Hello! 👋
-          </Link>
-        </div>
-      </Layout>
-    </footer>
-  );
-};
+        {">"} Hacked by Yasir Rahman © {new Date().getFullYear()}
+        <motion.span
+          className="inline-block ml-2"
+          animate={{ opacity: [1, 0, 1] }}
+          transition={{ duration: 1, repeat: Infinity }}
+        >
+          _
+        </motion.span>
+      </motion.p>
+    </div>
+  </footer>
+);
 
 export default Footer;
